First commit to change common enum with TileSize values into enum class, it works basically

+++ Project/FileBoard/IOBoard.cpp
@@ -108,7 +108,7 @@ void IOBoard::insertBoardValuesIntoStream( QDataStream& stream, Board& board )
 void IOBoard::insertBoardPicturesIntoStream( QDataStream& stream, BoardSize boardSize)
 {
     ImageProvider& provider = ImageProvider::getInstance();
-    stream << provider.getImageSquareSize( boardSize );
+    stream << (int) provider.getImageSquareSize( boardSize );
     vector< QImage* >& boardImages = provider.getImages( boardSize );
     int byteCountPerImage = boardImages.at( 0 )->byteCount();
     stream << byteCountPerImage;
